refactor(engine): Reexpport ExprContext in expressions root
