+++ tracecat/expressions/__init__.py
@@ -1,10 +1,11 @@
-from .engine import TemplateExpression
+from .engine import ExprContext, TemplateExpression
 from .eval import eval_templated_object, extract_templated_secrets
 from .validators import TemplateValidator
 
 __all__ = [
     "TemplateValidator",
     "TemplateExpression",
+    "ExprContext",
     "eval_templated_object",
     "extract_templated_secrets",
 ]
